% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_testfile.R
\name{create_testfile}
\alias{create_testfile}
\title{Use a spreadsheet to create files for use with the testthat package.}
\usage{
create_testfile(
  ss,
  sheet,
  manipulator = NULL,
  filename = NULL,
  filepath = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{ss}{dribble or filepath.}

\item{sheet}{character (vector). The names of the sheets to be read.}

\item{manipulator}{a function that manipulates the testdata once it has been read from the spreadsheet}

\item{filename}{character (vector). The default is testsheet_<sheet>.R. The number of given
filenames must match the length of testdata.}

\item{filepath}{character. The defallt is the current working directory}

\item{overwrite}{boolean. The default is FALSE.}
}
\description{
Use a spreadsheet to create files for use with the testthat package.
}
\examples{
ss_path <- testsheets_example("excel")
drop_param_y <- function(testdata) {
  testdata <- testdata \%>\% dplyr::select(-dplyr::starts_with("param_y"))
}
filename <- "testsheet_sum_x_only.R"
create_testfile(ss_path, sheet = "sum", manipulator = drop_param_y, filename=filename)
}
