% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_structures.R
\name{get_spreadsheet_from_drive}
\alias{get_spreadsheet_from_drive}
\title{Get spreadsheet from Google Drive}
\usage{
get_spreadsheet_from_drive(file_dribble)
}
\arguments{
\item{file_dribble}{a dribble containing a single googledrive file}
}
\value{
spreadsheet
}
\description{
Get spreadsheet from Google Drive
}
