% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_testsheet.R
\name{read_testsheet}
\alias{read_testsheet}
\title{Reading a testsheet into a testdata tibble}
\usage{
read_testsheet(ss, sheet, keep_user_columns = TRUE)
}
\arguments{
\item{ss}{dribble or filepath.}

\item{sheet}{character. The name of the sheet to be read.}

\item{keep_user_columns}{boolean. Keeps user type columns in testdata. Note that user
columns will still be removed upon writing testdata to a testfile.}
}
\value{
tibble containing testdata.
}
\description{
Reading a testsheet into a testdata tibble
}
\examples{
googledrive::drive_deauth()
googlesheets4::gs4_deauth()
ss <- testsheets_example("google sheets")
testdata <- read_testsheet(ss, "mean")
}
