% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_testsheet.R
\name{read_testsheets}
\alias{read_testsheets}
\title{Reading all testsheets in one file into a vector of testdata tibbles}
\usage{
read_testsheets(ss)
}
\arguments{
\item{ss}{dribble or filepath.}
}
\value{
a vector tibbles containing testdata for each sheet in a ss
}
\description{
Reading all testsheets in one file into a vector of testdata tibbles
}
\examples{
url <- testsheets_example("google sheets")
ss <- googledrive::drive_get(url)
testdata <- read_testsheets(ss)
}
