% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsheets_example.R
\name{testsheets_example}
\alias{testsheets_example}
\title{Get path to example file}
\usage{
testsheets_example(type = NULL)
}
\arguments{
\item{type}{Name of spreadsheet type. If `NULL`, the example file types will be listed.}
}
\value{
a path or url to the specified resource
}
\description{
Get path to example file
}
\examples{
testsheets_example()
testsheets_example("google sheets")
testsheets_example("excel")
}
