% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_testsheets.R
\name{use_testsheets}
\alias{use_testsheets}
\title{Set up `testsheet` infrastructure}
\usage{
use_testsheets()
}
\description{
Sets up data structures required to save a running list of testsheet source files.
}
