% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_testsheet.R
\name{read_testsheet}
\alias{read_testsheet}
\title{read_testsheets}
\usage{
read_testsheet(spreadsheet, sheet_name)
}
\arguments{
\item{spreadsheet}{A [dribble](https://googledrive.tidyverse.org/reference/dribble.html) that contains metadata for a single file.}

\item{sheet_name}{The name of the sheet, or "tab", in the spreadsheet.}
}
\value{
Tibble containing the contents of the testsheet.
}
\description{
read_testsheets
}
\examples{

}
